Started working on Agent and Target classes

+++ 000_Common/Body.h
@@ -11,7 +11,7 @@ public:
   Body(Vector2 position = Vector2{0,0},  float mass = 10.f, Color color = RED);
   
   void ApplyForce(Vector2 force);
-  void Update(uint32_t dt);
+  void Update(double dt);
   void Draw();
 
   Vector2 Attract(Body& body);
